requirements.txt
python-telegram-bot==21.6
aiohttp==3.10.11
python-dotenv==1.0.1
supabase==2.10.0
